+++ responses/tests/test_registries.py
@@ -5,7 +5,7 @@ from requests.exceptions import ConnectionError
 import responses
 from responses import registries
 from responses.registries import OrderedRegistry
-from responses.test_responses import assert_reset
+from responses.tests.test_responses import assert_reset
 
 
 def test_set_registry_not_empty():
